refactor: moved domain logic to data layer
