+++ app/src/main/java/com/example/gamitask/MainActivity.kt
@@ -5,7 +5,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.gamitask.features.taskmanagement.domain.viewmodel.TaskViewModel
+import com.example.gamitask.features.taskmanagement.data.viewmodel.TaskViewModel
 import com.example.gamitask.features.taskmanagement.presentation.components.ToDoApplication
 import com.example.gamitask.ui.theme.GamiTaskTheme
