+++ app/src/main/java/com/example/gamitask/features/taskmanagement/data/db/TaskDatabase.kt
@@ -2,7 +2,7 @@ package com.example.gamitask.features.taskmanagement.data.db
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import com.example.gamitask.features.taskmanagement.domain.models.Task
+import com.example.gamitask.features.taskmanagement.data.models.Task
 import com.example.gamitask.features.taskmanagement.data.dao.TaskDAO
 import android.content.Context
