--- app/src/main/java/com/example/gamitask/features/taskmanagement/domain/models/Task.kt~
@@ -1,11 +1,0 @@
-package com.example.gamitask.features.taskmanagement.domain.models
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import java.util.UUID
-
-@Entity (tableName = "tasks")
-// data class for task
-data class Task(
-    @PrimaryKey (autoGenerate = true) var id: Int=0,
-    val description: String,
-)
